Created MatchedUsersDB and methods, matched users no longer part of profile
