+++ src/main/java/com/lustermaniacs/companion/models/User.java
@@ -66,15 +66,4 @@ public class User {
     public void setProfile(Profile profile) {
         this.profile = profile;
     }
-
-    // Method to add a user to someone's sysmatched list, which is stored as list of usernames
-    public void addSysmatchedUser(String userName){
-        this.profile.addSysmatchedUser(userName);
-    }
-
-    // Method to remove a user from someone's sysmatched list
-    public void removeSysmatchedUser(String userName){
-        this.profile.removeSysmatchedUser(userName);
-    }
-
 }
